Add startup check for S3 write.
