+++ src/config.ts
@@ -1,5 +1,5 @@
 export const SERVICE_ACCOUNT_FILE = 'resources/client_credentials.json';
 export const EMAIL_INPUT_CSV = 'resources/user_emails.csv';
 export const MIGRATION_LOG_INPUT_CSV = 'resources/4 Users - TransactionItem.csv';
-export const OUTPUT_CSV = 'output/dataset.csv';
+export const OUTPUT_CSV = 'build-output/dataset.csv';
 export const SCOPES = ['https://www.googleapis.com/auth/drive'];
